+++ app/src/main/java/com/lintschool/communi/event/EventsBottmSheet.kt
@@ -1,10 +1,11 @@
-package com.lintschool.communi
+package com.lintschool.communi.event
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.lintschool.communi.data.Event
+import com.lintschool.communi.R
+import com.lintschool.communi.event.data.Event
 
 class EventsBottmSheet : BottomSheetDialogFragment() {
     private var suggestionEventList: MutableList<Event> = mutableListOf()
