+++ app/src/main/java/com/lintschool/communi/event/adapters/AroundEventAdapter.kt
@@ -1,11 +1,11 @@
-package com.lintschool.communi.adapters
+package com.lintschool.communi.event.adapters
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.lintschool.communi.R
-import com.lintschool.communi.data.Event
+import com.lintschool.communi.event.data.Event
 import kotlinx.android.synthetic.main.item_arround_event.view.*
 
 class AroundEventAdapter(
@@ -21,7 +21,7 @@ class AroundEventAdapter(
         parent: ViewGroup,
         viewType: Int
     ):
-        ArroundEventViewHolder {
+            ArroundEventViewHolder {
             val itemView = LayoutInflater.from(parent.context)
                 .inflate(R.layout.item_arround_event, parent, false)
             return ArroundEventViewHolder(itemView)
