+++ app/src/main/java/com/lintschool/communi/event/data/Event.kt
@@ -1,4 +1,4 @@
-package com.lintschool.communi.data
+package com.lintschool.communi.event.data
 
 data class Event(
     val name: String = "",
