+++ app/src/main/java/com/lintschool/communi/event/fragments/FeedFragment.kt
@@ -1,4 +1,4 @@
-package com.lintschool.communi.fragments
+package com.lintschool.communi.event.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.lintschool.communi.R
 
-class GuestListFragment : Fragment() {
+class FeedFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        inflater.inflate(R.layout.fragment_guest_list, container, false)
+        inflater.inflate(R.layout.fragment_feed, container, false)
 }
