+++ app/src/main/res/layout/activity_bottom_sheet_content.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".Main">
+    tools:context=".event.Main">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/textView"
