+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:animateLayoutChanges="true"
-    tools:context=".MainActivity">
+    tools:context=".event.MainActivity">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
