+++ app/src/main/res/layout/activity_main_.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main">
+    tools:context=".event.Main">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
